+++ spec/tests/constants_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 # *******************************************************************************
-# OpenStudio(R), Copyright (c) 2008-2021, Alliance for Sustainable Energy, LLC.
-# BuildingSync(R), Copyright (c) 2015-2021, Alliance for Sustainable Energy, LLC.
+# OpenStudio(R), Copyright (c) 2008-2022, Alliance for Sustainable Energy, LLC.
+# BuildingSync(R), Copyright (c) 2015-2022, Alliance for Sustainable Energy, LLC.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -45,6 +45,9 @@ RSpec.describe 'BuildingSync constants' do
   end
   it 'should have a SCHEMA_2_2_0_URL' do
     expect(SCHEMA_2_2_0_URL).to eq('https://raw.githubusercontent.com/BuildingSync/schema/v2.2.0/BuildingSync.xsd')
+  end
+  it 'should have a SCHEMA_2_4_0_URL' do
+    expect(SCHEMA_2_4_0_URL).to eq('https://raw.githubusercontent.com/BuildingSync/schema/v2.4.0/BuildingSync.xsd')
   end
   it 'should have a PHASE_0_BASE_OSW_FILE_PATH and the file should exist' do
     expect(File.exist?(PHASE_0_BASE_OSW_FILE_PATH)).to be true