+++ spec/tests/translator_sizing_run_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 # *******************************************************************************
-# OpenStudio(R), Copyright (c) 2008-2021, Alliance for Sustainable Energy, LLC.
-# BuildingSync(R), Copyright (c) 2015-2021, Alliance for Sustainable Energy, LLC.
+# OpenStudio(R), Copyright (c) 2008-2022, Alliance for Sustainable Energy, LLC.
+# BuildingSync(R), Copyright (c) 2015-2022, Alliance for Sustainable Energy, LLC.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -43,23 +43,23 @@ RSpec.describe 'BuildingSync' do
     tests_to_run = [
       # file_name, standard, epw_path, schema_version
       # Building 151
-      ['building_151.xml', CA_TITLE24, nil, 'v2.2.0'],
-      ['building_151.xml', ASHRAE90_1, nil, 'v2.2.0'],
-      ['building_151.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
-      ['building_151.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
+      ['building_151.xml', CA_TITLE24, nil, 'v2.4.0'],
+      ['building_151.xml', ASHRAE90_1, nil, 'v2.4.0'],
+      ['building_151.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
+      ['building_151.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
 
       # Building 151 n1
-      ['building_151_n1.xml', CA_TITLE24, nil, 'v2.2.0'],
-      ['building_151_n1.xml', ASHRAE90_1, nil, 'v2.2.0'],
-      ['building_151_n1.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
-      ['building_151_n1.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
+      ['building_151_n1.xml', CA_TITLE24, nil, 'v2.4.0'],
+      ['building_151_n1.xml', ASHRAE90_1, nil, 'v2.4.0'],
+      ['building_151_n1.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
+      ['building_151_n1.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
 
       # DC GSA Headquarters
-      ['DC GSA Headquarters.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
+      ['DC GSA Headquarters.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
 
       # DC GSA Headquarters with Climate Zone
-      ['DC GSA HeadquartersWithClimateZone.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
-      ['DC GSA HeadquartersWithClimateZone.xml', ASHRAE90_1, nil, 'v2.2.0'],
+      ['DC GSA HeadquartersWithClimateZone.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
+      ['DC GSA HeadquartersWithClimateZone.xml', ASHRAE90_1, nil, 'v2.4.0'],
 
       # L100 Audit
       # None working
@@ -68,27 +68,27 @@ RSpec.describe 'BuildingSync' do
       # None of these should work, see errors that get caught in next section.
 
       # L000_OpenStudio_Pre-Simulation-01
-      ['L000_OpenStudio_Pre-Simulation_01.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
+      ['L000_OpenStudio_Pre-Simulation_01.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
 
       # L000_OpenStudio_Pre-Simulation-02
-      ['L000_OpenStudio_Pre-Simulation_02.xml', ASHRAE90_1, nil, 'v2.2.0'],
-      ['L000_OpenStudio_Pre-Simulation_02.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
-      ['L000_OpenStudio_Pre-Simulation_02.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
+      ['L000_OpenStudio_Pre-Simulation_02.xml', ASHRAE90_1, nil, 'v2.4.0'],
+      ['L000_OpenStudio_Pre-Simulation_02.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
+      ['L000_OpenStudio_Pre-Simulation_02.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
 
       # L000_OpenStudio_Pre-Simulation-03
-      ['L000_OpenStudio_Pre-Simulation_03.xml', ASHRAE90_1, nil, 'v2.2.0'],
-      ['L000_OpenStudio_Pre-Simulation_03.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
-      ['L000_OpenStudio_Pre-Simulation_03.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
+      ['L000_OpenStudio_Pre-Simulation_03.xml', ASHRAE90_1, nil, 'v2.4.0'],
+      ['L000_OpenStudio_Pre-Simulation_03.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
+      ['L000_OpenStudio_Pre-Simulation_03.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
 
       # L000_OpenStudio_Pre-Simulation-04
-      ['L000_OpenStudio_Pre-Simulation_04.xml', ASHRAE90_1, nil, 'v2.2.0'],
-      ['L000_OpenStudio_Pre-Simulation_04.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
-      ['L000_OpenStudio_Pre-Simulation_04.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
+      ['L000_OpenStudio_Pre-Simulation_04.xml', ASHRAE90_1, nil, 'v2.4.0'],
+      ['L000_OpenStudio_Pre-Simulation_04.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
+      ['L000_OpenStudio_Pre-Simulation_04.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
 
       # Office_Carolina
-      ['Office_Carolina.xml', ASHRAE90_1, nil, 'v2.2.0'],
-      ['Office_Carolina.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0'],
-      ['Office_Carolina.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0']
+      ['Office_Carolina.xml', ASHRAE90_1, nil, 'v2.4.0'],
+      ['Office_Carolina.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0'],
+      ['Office_Carolina.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0']
     ]
     tests_to_run.each do |test|
       it "File: #{test[0]}. Standard: #{test[1]}. EPW_Path: #{test[2]}. File Schema Version: #{test[3]}" do
@@ -104,70 +104,70 @@ RSpec.describe 'BuildingSync' do
 
       #####################################
       ## building_151_level1
-      ['building_151_level1.xml', CA_TITLE24, nil, 'v2.2.0', "undefined method `add_internal_loads' for nil:NilClass"],
-      ['building_151_level1.xml', ASHRAE90_1, nil, 'v2.2.0', "undefined method `add_internal_loads' for nil:NilClass"],
-      ['building_151_level1.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', "undefined method `add_internal_loads' for nil:NilClass"],
-      ['building_151_level1.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', "undefined method `add_internal_loads' for nil:NilClass"],
+      ['building_151_level1.xml', CA_TITLE24, nil, 'v2.4.0', "undefined method `add_internal_loads' for nil:NilClass"],
+      ['building_151_level1.xml', ASHRAE90_1, nil, 'v2.4.0', "undefined method `add_internal_loads' for nil:NilClass"],
+      ['building_151_level1.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', "undefined method `add_internal_loads' for nil:NilClass"],
+      ['building_151_level1.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', "undefined method `add_internal_loads' for nil:NilClass"],
 
       #####################################
       ## BuildingSync Website Valid Schema
-      ['BuildingSync Website Valid Schema.xml', CA_TITLE24, nil, 'v2.2.0', 'Building ID: Building001. OccupancyClassification must be defined at either the Site or Building level.'],
-      ['BuildingSync Website Valid Schema.xml', ASHRAE90_1, nil, 'v2.2.0', 'Building ID: Building001. OccupancyClassification must be defined at either the Site or Building level.'],
-      ['BuildingSync Website Valid Schema.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', 'Building ID: Building001. OccupancyClassification must be defined at either the Site or Building level.'],
-      ['BuildingSync Website Valid Schema.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', 'Building ID: Building001. OccupancyClassification must be defined at either the Site or Building level.'],
+      ['BuildingSync Website Valid Schema.xml', CA_TITLE24, nil, 'v2.4.0', 'Building ID: Building001. OccupancyClassification must be defined at either the Site or Building level.'],
+      ['BuildingSync Website Valid Schema.xml', ASHRAE90_1, nil, 'v2.4.0', 'Building ID: Building001. OccupancyClassification must be defined at either the Site or Building level.'],
+      ['BuildingSync Website Valid Schema.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', 'Building ID: Building001. OccupancyClassification must be defined at either the Site or Building level.'],
+      ['BuildingSync Website Valid Schema.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', 'Building ID: Building001. OccupancyClassification must be defined at either the Site or Building level.'],
 
       # #####################################
       # ## DC GSA Headquarters
       # Really this fails because: [BuildingSync.GetBCLWeatherFile.download_weather_file_from_city_name] <1> Error, could not find uid for state DC and city Washington. Initial count of weather files: 10. Please try a different weather file.
-      ['DC GSA Headquarters.xml', ASHRAE90_1, nil, 'v2.2.0', 'BuildingSync.Building.set_weather_and_climate_zone: epw_file_path is false: false'],
-      ['DC GSA Headquarters.xml', CA_TITLE24, nil, 'v2.2.0', 'Did not find a', false],
-      ['DC GSA Headquarters.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', 'Did not find a', false],
+      ['DC GSA Headquarters.xml', ASHRAE90_1, nil, 'v2.4.0', 'BuildingSync.Building.set_weather_and_climate_zone: epw_file_path is false: false'],
+      ['DC GSA Headquarters.xml', CA_TITLE24, nil, 'v2.4.0', 'Did not find a', false],
+      ['DC GSA Headquarters.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', 'Did not find a', false],
 
       ####################################
       # DC GSA HeadquartersWithClimateZone
-      ['DC GSA HeadquartersWithClimateZone.xml', CA_TITLE24, nil, 'v2.2.0', 'Did not find a', false],
-      ['DC GSA HeadquartersWithClimateZone.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', 'Did not find a', false],
+      ['DC GSA HeadquartersWithClimateZone.xml', CA_TITLE24, nil, 'v2.4.0', 'Did not find a', false],
+      ['DC GSA HeadquartersWithClimateZone.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', 'Did not find a', false],
 
       # #####################################
       # ## L100 Audit
       # Trace/BPT trap: 5 gets hit for following 2 lines
-      # ['L100_Audit.xml', CA_TITLE24, nil, 'v2.2.0', "Error, cannot find local component for: 1ed4ea50-edc6-0131-1b8b-48e0eb16a403.  Please try a different weather file."],
-      # ['L100_Audit.xml', ASHRAE90_1, nil, 'v2.2.0', "Error, cannot find local component for: 1ed4ea50-edc6-0131-1b8b-48e0eb16a403.  Please try a different weather file."],
-      ['L100_Audit.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', "undefined method `add_internal_loads' for nil:NilClass"],
-      ['L100_Audit.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', "undefined method `add_internal_loads' for nil:NilClass"],
+      # ['L100_Audit.xml', CA_TITLE24, nil, 'v2.4.0', "Error, cannot find local component for: 1ed4ea50-edc6-0131-1b8b-48e0eb16a403.  Please try a different weather file."],
+      # ['L100_Audit.xml', ASHRAE90_1, nil, 'v2.4.0', "Error, cannot find local component for: 1ed4ea50-edc6-0131-1b8b-48e0eb16a403.  Please try a different weather file."],
+      ['L100_Audit.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', "undefined method `add_internal_loads' for nil:NilClass"],
+      ['L100_Audit.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', "undefined method `add_internal_loads' for nil:NilClass"],
 
       # #####################################
       # ## Golden File
       # Trace/BPT trap: 5 gets hit for following 2 lines
-      # ['Golden Test File.xml', CA_TITLE24, nil, 'v2.2.0', "Error, cannot find local component for: fa8c9ff0-edc4-0131-a9f8-48e0eb16a403.  Please try a different weather file."],
-      # ['Golden Test File.xml', ASHRAE90_1, nil, 'v2.2.0', "Error, cannot find local component for: fa8c9ff0-edc4-0131-a9f8-48e0eb16a403.  Please try a different weather file."],
-      ['Golden Test File.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', "undefined method `add_internal_loads' for nil:NilClass"],
-      ['Golden Test File.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', "undefined method `add_internal_loads' for nil:NilClass"],
+      # ['Golden Test File.xml', CA_TITLE24, nil, 'v2.4.0', "Error, cannot find local component for: fa8c9ff0-edc4-0131-a9f8-48e0eb16a403.  Please try a different weather file."],
+      # ['Golden Test File.xml', ASHRAE90_1, nil, 'v2.4.0', "Error, cannot find local component for: fa8c9ff0-edc4-0131-a9f8-48e0eb16a403.  Please try a different weather file."],
+      ['Golden Test File.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', "undefined method `add_internal_loads' for nil:NilClass"],
+      ['Golden Test File.xml', ASHRAE90_1, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', "undefined method `add_internal_loads' for nil:NilClass"],
 
       # #####################################
       # L000_OpenStudio_Pre-Simulation-01
       # TODO: Fix download BCL function or figure out workaround
       # Trace/BPT trap: 5 gets hit for following 2 lines
-      # ['L000_OpenStudio_Pre-Simulation_01.xml', CA_TITLE24, nil, 'v2.2.0', "Error, cannot find local component for: 1fd3d630-edc5-0131-b802-48e0eb16a403.  Please try a different weather file."],
-      # ['L000_OpenStudio_Pre-Simulation_01.xml', ASHRAE90_1, nil, 'v2.2.0', "Error, cannot find local component for: 1fd3d630-edc5-0131-b802-48e0eb16a403.  Please try a different weather file."],
+      # ['L000_OpenStudio_Pre-Simulation_01.xml', CA_TITLE24, nil, 'v2.4.0', "Error, cannot find local component for: 1fd3d630-edc5-0131-b802-48e0eb16a403.  Please try a different weather file."],
+      # ['L000_OpenStudio_Pre-Simulation_01.xml', ASHRAE90_1, nil, 'v2.4.0', "Error, cannot find local component for: 1fd3d630-edc5-0131-b802-48e0eb16a403.  Please try a different weather file."],
 
       # We have issues with old CBES files
-      ['L000_OpenStudio_Pre-Simulation_01.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.2.0', 'Did not find a', false],
+      ['L000_OpenStudio_Pre-Simulation_01.xml', CA_TITLE24, File.join(SPEC_WEATHER_DIR, 'USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw'), 'v2.4.0', 'Did not find a', false],
 
       # #####################################
       # L000_OpenStudio_Pre-Simulation-02
       # Really this fails because: Could not determine the weather file for climate zone: CEC T24-CEC6A.
-      ['L000_OpenStudio_Pre-Simulation_02.xml', CA_TITLE24, nil, 'v2.2.0', 'no implicit conversion of nil into String'],
+      ['L000_OpenStudio_Pre-Simulation_02.xml', CA_TITLE24, nil, 'v2.4.0', 'no implicit conversion of nil into String'],
 
       # #####################################
       # L000_OpenStudio_Pre-Simulation-03
       # Really this fails because: Could not determine the weather file for climate zone: CEC T24-CEC1A.
-      ['L000_OpenStudio_Pre-Simulation_03.xml', CA_TITLE24, nil, 'v2.2.0', 'no implicit conversion of nil into String'],
+      ['L000_OpenStudio_Pre-Simulation_03.xml', CA_TITLE24, nil, 'v2.4.0', 'no implicit conversion of nil into String'],
 
       # #####################################
       # L000_OpenStudio_Pre-Simulation-04
       # Really this fails because: Could not determine the weather file for climate zone: CEC T24-CEC6A.
-      ['L000_OpenStudio_Pre-Simulation_04.xml', CA_TITLE24, nil, 'v2.2.0', 'no implicit conversion of nil into String'],
+      ['L000_OpenStudio_Pre-Simulation_04.xml', CA_TITLE24, nil, 'v2.4.0', 'no implicit conversion of nil into String'],
 
       #####################################
       ## AT_example_property_report_25
@@ -186,7 +186,7 @@ RSpec.describe 'BuildingSync' do
       #####################################
       ## Office Carolina
       # Really this fails because: Could not determine the weather file for climate zone: CEC T24-CEC6A.
-      ['Office_Carolina.xml', CA_TITLE24, nil, 'v2.2.0', 'no implicit conversion of nil into String']
+      ['Office_Carolina.xml', CA_TITLE24, nil, 'v2.4.0', 'no implicit conversion of nil into String']
     ]
     tests_to_run.each do |test|
       it "Should fail with message: #{test[4]}" do